requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
futures==3.1.1
goslate==1.5.1
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
oauthlib==3.1.1
pycparser==2.20
PyDictionary==2.0.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
